♻️ Refactor : Reorganize directory structure for user
