+++ src/main/java/com/run_us/server/domains/running/service/RunningResultService.java
@@ -8,7 +8,7 @@ import com.run_us.server.domains.running.exceptions.RunningErrorCode;
 import com.run_us.server.domains.running.exceptions.RunningException;
 import com.run_us.server.domains.running.repository.PersonalRecordRepository;
 import com.run_us.server.domains.running.repository.RunningRepository;
-import com.run_us.server.domains.user.model.User;
+import com.run_us.server.domains.user.domain.User;
 import com.run_us.server.domains.user.repository.UserRepository;
 import jakarta.transaction.Transactional;
 import java.util.List;
