+++ src/main/java/com/run_us/server/domains/user/domain/Gender.java
@@ -1,4 +1,4 @@
-package com.run_us.server.domains.user.model;
+package com.run_us.server.domains.user.domain;
 
 public enum Gender {
   FEMALE,
