+++ src/main/java/com/run_us/server/domains/user/domain/SocialProvider.java
@@ -1,0 +1,5 @@
+package com.run_us.server.domains.user.domain;
+
+public enum SocialProvider {
+  KAKAO
+}
