+++ src/main/java/com/run_us/server/domains/user/domain/User.java
@@ -1,4 +1,4 @@
-package com.run_us.server.domains.user.model;
+package com.run_us.server.domains.user.domain;
 
 import com.run_us.server.global.common.DateAudit;
 import io.hypersistence.tsid.TSID;
