--- src/main/java/com/run_us/server/domains/user/model/SocialProvider.java
@@ -1,5 +1,0 @@
-package com.run_us.server.domains.user.model;
-
-public enum SocialProvider {
-  KAKAO
-}
