+++ src/test/java/com/run_us/server/domains/running/RunningControllerTest.java
@@ -5,9 +5,8 @@ import com.run_us.server.domains.running.controller.JoinedParticipantsDto;
 import com.run_us.server.domains.running.controller.RunningController;
 import com.run_us.server.domains.running.domain.Running;
 import com.run_us.server.domains.running.repository.RunningRepository;
-import com.run_us.server.domains.running.service.RunningPreparationService;
-import com.run_us.server.domains.user.model.User;
-import com.run_us.server.domains.user.model.UserFixtures;
+import com.run_us.server.domains.user.domain.User;
+import com.run_us.server.domains.user.domain.UserFixtures;
 import com.run_us.server.domains.user.repository.UserRepository;
 import com.run_us.server.global.common.SuccessResponse;
 import java.util.List;
