+++ src/test/java/com/run_us/server/domains/running/RunningFixtures.java
@@ -3,7 +3,7 @@ package com.run_us.server.domains.running;
 import com.run_us.server.domains.running.domain.Running;
 import com.run_us.server.domains.running.domain.RunningConstraints;
 import com.run_us.server.domains.running.domain.RunningDescription;
-import com.run_us.server.domains.user.model.User;
+import com.run_us.server.domains.user.domain.User;
 import com.run_us.server.global.utils.PointGenerator;
 import java.util.List;
 import org.locationtech.jts.geom.Point;
