+++ src/test/java/com/run_us/server/domains/running/RunningTest.java
@@ -5,8 +5,8 @@ import static org.junit.jupiter.api.Assertions.*;
 import com.run_us.server.domains.running.domain.Running;
 import com.run_us.server.domains.running.domain.RunningConstraints;
 import com.run_us.server.domains.running.domain.RunningDescription;
-import com.run_us.server.domains.user.model.User;
-import com.run_us.server.domains.user.model.UserFixtures;
+import com.run_us.server.domains.user.domain.User;
+import com.run_us.server.domains.user.domain.UserFixtures;
 import com.run_us.server.global.utils.PointGenerator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
