+++ src/test/java/com/run_us/server/domains/user/domain/UserFixtures.java
@@ -1,4 +1,4 @@
-package com.run_us.server.domains.user.model;
+package com.run_us.server.domains.user.domain;
 
 import java.time.LocalDate;
